fix razor css files, now they works

+++ HITSBlazor/Models/Ideas/Entities/Idea.cs
@@ -27,7 +27,7 @@ namespace HITSBlazor.Models.Ideas.Entities
 
         public int Suitability { get; set; }
         public int Budget { get; set; }
-        public int PreAssessment { get; set; }
+        public double PreAssessment { get; set; }
         public double? Rating { get; set; }
 
         public bool IsChecked { get; set; }
